+++ src/kubernetes/resources/data/event.rs
@@ -32,7 +32,7 @@ fn data_filtered(object: &DynamicObject) -> ResourceData {
 }
 
 pub fn header_filtered() -> Header {
-    let mut header = Header::from(
+    Header::from(
         NAMESPACE,
         Some(Box::new([
             Column::fixed("COUNT", 6, true),
@@ -40,9 +40,8 @@ pub fn header_filtered() -> Header {
             Column::bound("MESSAGE", 15, 150, false),
         ])),
         Rc::new([' ', 'N', 'C', 'T', 'M', 'A']),
-    );
-    header.set_sort_info(5, false);
-    header
+    )
+    .with_sort_info(5, false)
 }
 
 fn data_full(object: &DynamicObject) -> ResourceData {
@@ -73,20 +72,16 @@ fn data_full(object: &DynamicObject) -> ResourceData {
 }
 
 pub fn header_full() -> Header {
-    let mut last = Column::fixed("LAST", 6, true);
-    last.has_reversed_order = true;
-
-    let mut header = Header::from(
+    Header::from(
         NAMESPACE,
         Some(Box::new([
-            last,
+            Column::fixed("LAST", 6, true).with_reversed_order(),
             Column::fixed("COUNT", 6, true),
             Column::bound("TYPE", 6, 7, false),
             Column::bound("REASON", 6, 25, false),
             Column::bound("OBJECT", 15, 70, false),
         ])),
         Rc::new([' ', 'N', 'L', 'C', 'T', 'R', 'O', 'A']),
-    );
-    header.set_sort_info(2, false);
-    header
+    )
+    .with_sort_info(2, false)
 }